requirements.txt
streamlit 
gspread 
oauth2client 
requests 
python-dotenv 
pandas 
openai 
pyngrok 
requests 
groq